fix fonctionnalité admin flux + gestion affichage header

+++ src/Controller/FluxController.php
@@ -20,13 +20,51 @@ final class FluxController extends AbstractController
     #[Route(name: 'app_flux_index', methods: ['GET'])]
     public function index(FluxRepository $fluxRepository): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_home');
+        }
+        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_home');
+        }
         return $this->render('flux/index.html.twig', [
             'fluxes' => $fluxRepository->findAll(),
+        ]);
+    }
+
+    #[Route('/newFlux', name: 'app_flux_newFlux', methods: ['GET', 'POST'])]
+    public function newFlux(Request $request, EntityManagerInterface $entityManager): Response
+    {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_home');
+        }
+        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_home');
+        }
+        $flux = new Flux();
+        $form = $this->createForm(FluxType::class, $flux);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($flux);
+            $entityManager->flush();
+
+            return $this->redirectToRoute('app_flux_index', [], Response::HTTP_SEE_OTHER);
+        }
+
+        return $this->render('flux/new.html.twig', [
+            'flux' => $flux,
+            'form' => $form,
         ]);
     }
     #[Route('/new', name: 'app_flux_new', methods: ['POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_home');
+        }
+        if ($this->getUser()->getRoles()) {
+            return $this->redirectToRoute('app_home');
+        }
         $titre = $request->request->get('titre');
         $message = $request->request->get('message');
         $refFluxId = $request->request->get('refFluxId');
@@ -82,7 +120,7 @@ final class FluxController extends AbstractController
         if ($this->getUser() == null) {
             return $this->redirectToRoute('app_home');
         }
-        if ($flux->getRole()[0] != $this->getUser()->getRoles()[0] && $flux->getId() != 1) {
+        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
             return $this->redirectToRoute('app_home');
         }
         $search = $request->query->get('search', null);
@@ -115,6 +153,12 @@ final class FluxController extends AbstractController
     #[Route('/{id}/edit', name: 'app_flux_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Flux $flux, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_home');
+        }
+        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_home');
+        }
 
         $form = $this->createForm(FluxType::class, $flux);
         $form->handleRequest($request);
@@ -134,6 +178,12 @@ final class FluxController extends AbstractController
     #[Route('/{id}', name: 'app_flux_delete', methods: ['POST'])]
     public function delete(Request $request, Flux $flux, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser() == null) {
+            return $this->redirectToRoute('app_home');
+        }
+        if (!in_array('ROLE_ADMIN', $this->getUser()->getRoles())) {
+            return $this->redirectToRoute('app_home');
+        }
         if ($this->isCsrfTokenValid('delete'.$flux->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($flux);
             $entityManager->flush();
